feat: #12 일부 클래스명 변경
