+++ 1_CodeSoruce/src/main/java/com/bughunters/mountainspirit/crew/command/controller/CrewCommendController.java
@@ -3,29 +3,23 @@ package com.bughunters.mountainspirit.crew.command.controller;
 import com.bughunters.mountainspirit.crew.command.dto.CrewDeleteDTO;
 import com.bughunters.mountainspirit.crew.command.dto.CrewModifyDTO;
 import com.bughunters.mountainspirit.crew.command.dto.CrewRegistDTO;
-import com.bughunters.mountainspirit.crew.command.entity.Crew;
 import com.bughunters.mountainspirit.crew.command.service.CrewCommendService;
-import com.bughunters.mountainspirit.crew.command.service.CrewCommendServiceInter;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.coyote.Request;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
-import java.time.LocalDateTime;
 
 @RestController
 @RequestMapping("/crew")
 @Slf4j
 public class CrewCommendController {
-    private CrewCommendServiceInter crewCommendServiceInter;
+    private CrewCommendService crewCommendService;
 
     @Autowired
-    public CrewCommendController(CrewCommendServiceInter crewCommendServiceInter) {
-        this.crewCommendServiceInter = crewCommendServiceInter;
+    public CrewCommendController(CrewCommendService crewCommendService) {
+        this.crewCommendService = crewCommendService;
     }
 
     @GetMapping("/health")
@@ -35,19 +29,19 @@ public class CrewCommendController {
 
     @PostMapping("/crew-create")
     public ResponseEntity<?> createCrew(@RequestBody CrewRegistDTO crewRegistDTO) {
-        crewCommendServiceInter.insertCrew(crewRegistDTO);
+        crewCommendService.insertCrew(crewRegistDTO);
         return ResponseEntity.created(URI.create("crew/crew_info")).build();
     }
 
     @PutMapping("/crew-modify")
     public ResponseEntity<?> modifyCrew(@RequestBody CrewModifyDTO crewModifyDTO) {
-        crewCommendServiceInter.modifyCrew(crewModifyDTO);
+        crewCommendService.modifyCrew(crewModifyDTO);
         return ResponseEntity.created(URI.create("crew/crew_info/" + crewModifyDTO.getId())).build();
     }
 
     @PutMapping("/crew-delete")
     public ResponseEntity<?> deleteCrew(@RequestBody CrewDeleteDTO crewDeleteDTO) {
-        crewCommendServiceInter.deleteCrew(crewDeleteDTO);
+        crewCommendService.deleteCrew(crewDeleteDTO);
         return ResponseEntity.noContent().build();
     }
 }
