--- 1_CodeSoruce/src/main/java/com/bughunters/mountainspirit/crew/query/service/CrewQueryServiceInter.java
@@ -1,11 +1,0 @@
-package com.bughunters.mountainspirit.crew.query.service;
-
-import com.bughunters.mountainspirit.crew.query.dto.CrewDTO;
-
-import java.util.List;
-
-public interface CrewQueryServiceInter {
-    List<CrewDTO> findAllCrews();
-
-    CrewDTO findOneCrewById(Long crewId);
-}
